ai-chat/requirements.txt
transformers==4.36.2
torch==2.2.0
datasets==2.16.1
accelerate==0.27.2
sentencepiece==0.1.99
gradio==4.19.2
python-dotenv==1.0.1 